Add setId to MorphiaModel
This method is needed when you have a form and the id is a PathParam (for example). In this case your model has a null objectId and you have to set manually the Id to update using `mongoDB.getDatastore().save(myModel);`, if don't, Morphia will create another entry instead of update the older entry.